requirements.txt
<<<<<<< HEAD
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
pydantic
streamlit
requests
=======
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
pydantic
streamlit
requests
>>>>>>> 0cb9a91 (Added FastAPI backend)
